msp/cloudrun: rename Cloud Run resources to include environment and location (#59848)
As part of https://github.com/sourcegraph/managed-services/issues/186 we're leaning towards leveraging Cloud Deploy to establish deployment stages and pipelines. Sadly, Cloud Deploy targets for Cloud Run require that the Cloud Run service name to be unique across environments - this requires renaming all MSP Cloud Run services to include the environment ID, which is a destructive change. In anticipation of [multi-region deployments in the future](https://cloud.google.com/run/docs/multiple-regions#deploy-service), this change also adds the location to the name while we're at it.

While testing this, it turns out a variety of our network resources cannot accommodate a Cloud Run recreation in Terraform, so I've renamed the affected components to use the Cloud Run service name instead so that they get recreated entirely when the Cloud Run service is recreated via a name change. I also noticed that the external uptime alert might not be firing aggressively enough to catch the downtime - I've adjusted the thresholds here, and also aggregated it so that we just get 1 alert (closes https://github.com/sourcegraph/managed-services/issues/393)

**This is a downtime-incurring change**, but the downtime should be minimal, as core components that require propagation like external addresses remain unchanged.

## Test plan

Deployed to `msp-tesbed` `robert` and `telemetry-gateway` `dev` environment