build: Merge sg-syntax into scip-syntax
The bulk of the logic in the sg-syntax crate is related to highlighting.
Since that is generic information related to highlighting, move
it to scip-syntax which will be renamed to syntax-analysis in the future.

Moving it inside the syntax-highlighter itself was another option,
but I didn't go for that, so that we can separately benchmark the
syntax-analysis crate, instead of introducing a new library target
in syntax-highlighter and then benchmarking that.