search: encode streaming SSE with gzip
We have noticed large response payloads on dotcom which are not gzipped
compressed. This should help make a dent in the amount of data a client
needs to download over the wire.

Alternatively I considered upgrading all uses of StreamWriter to support
gzip. I was surpised but happy to see StreamWriter now has multiple
consumers outside of search (completions and blame). However, this was
relatively complicated so targetting just our main streaming endpoint.

Test Plan: Added unit tests. Additionally debugged why the middleware
wasn't working before and how to make it work with
https://gist.github.com/keegancsmith/4e133d976f7042f0b56b6f1f062dae7b