svelte: Fix search results rendering issues (#59806)
This commit fixes a couple of issues I noticed with search results
rendering:

It was possible that we were rendering fewer plaintext lines than
highlighted lines, which caused a jump as soon as the highlighted
version was available. That's because we might show a truncated version
of the chunk, but we'll always fetch the highlighting information for
the full chunk. We need to account for that by truncating the
highlighted lines too.

There have been several issues when search result components have been
reused for different search results. Reuse can happen when navigating
between cached results via the browser back/forward buttons. There are
multiple ways to fix this. I choose to change the components themselves
so that any state would be updated/recomputed when the result changes.
In particular `FileContentSearchResult` will now refetch highlighting
information when its input changes.
Using `{#await ...}` inside the loop doesn't seem ideal but I didn't
notice any problems with it.

There have been some inconsistencies with capturing/restoring the scroll
position of the search results list. I switched to using SvelteKit's
native snapshot feature instead.