search: Disable 'jump to suggestion' behavior in secondary query inputs (#59941)
The "old" search input is still used in places other than the search and repository pages (i.e. everywhere that's not the main search input). These places are solely used for entering a query, not for navigating around, and so "jump to suggestion" for repositories and files should be disabled.