search: boost phrase queries (#59940)
This uses the new query.Boost operator exposed by Zoekt to boost phrase queries. The feature is behind the feature-flag "search-boost-phrase".

The heuristic to detect phrases is very simple: We detect phrases if the query is a top-level AND query without nesting or negated terms.

Test plan:

New unit test
Manual testing: I checked locally with the feature flag turned off and on and compared the results.