gitserver: Simplify accesslogging
This PR reduces the number of interceptors that need to be invoked to check if we should be logging.
We only log if accesslog has been called in the handler.
This also makes it so that we cannot forget to add the method here so that access logging works.
That was not the case for the perforce handlers before.

## Test plan

Ran manually and checked that access logs are logged correctly.