search: use gzip.BestSpeed for streaming endpoint (#59965)
We recently forced gzip compression on responses from the streaming
endpoint. However, we have noticed a significant performance regression
in search blitz which is likely due to it. Search blitz runs in the same
data-center as dotcom, so the compression didn't help download speed but
did vastly increase response time.

Camden pointed out using gzip.BestSpeed on the original PR and in my
testing it is a significant improvement on CPU use with an acceptable
decrease in compression ratio. Here is the table from my testing on a
real life example:

  | encoding   |      size | time  |
  |------------+-----------+-------|
  | none       | 287728396 | 0.1s  |
  | zstd       |  64662426 | 0.9s  |
  | gzip-speed | 102231337 | 2.4s  |
  | br         |  61296159 | 9.1s  |
  | gzip       |  86795072 | 10.7s |
  | deflate    |  86795060 | 10.7s |

Test Plan: existing unit tests cover correctly handling compression.
Will monitor impact on our search-blitz when this rolls out.