Search preview panel: use static highlight ranges from search results (#59715)
Not all searches can be represented as a single regex pattern, so we can't easily translate a search query into a code mirror search that shows the same results. However, we already have the set of matched ranges from our search results, so we can use that to provide a static set of highlights in the preview panel.