search: respect whitespace for phrase boost (#59974)
Terms in phrases might be separated by more than one whitespace. With this change we respect the original query when we construct the phrase query. Before, we always joined terms with " ".  

Additionally we couple this feature to keyword search instead of having it as a feature flag.

Note: The transformation now happens during the planning stage instead of the executing stage. I think this is a more obvious spot as it happens next to other transformations on the query plan.

Test plan:
updated unit test