search: revert to single space phrase boosting
During manual testing we noticed that queries that include reserved
keywords, such as `(foo and bar) and bas` led to undesired phrase
queries. For the case of `(foo and bar) and bas` we would search the
phrase `foo and bar) and bas`. Hence we revert the logic to a much
simpler join by " ".

I also fix another bug which was caused by missing parser flags.
Essentially, the missing flags led to parser errors if specials
characters were used in the phrase.

Test plan:
updated unit test