spike: display math expressions with MathJax
When rendering Markdown (and Jupyter) files
we can load MathJax scripts on the fly and
have them render LaTeX/MathML.

Spike because I'm not sure if loading external scripts is
a good-enough solution. Works quite well though.
Doing this server-side does not seem feasible,
and I haven't found any reliable Go libraries to do it.
MathJax is something of an industry standard and the
go-to solution here.

NOTE: Markdown.tsx is used in a lot of places, but
I've added a parameter to only enable MathJax for "rendered file"
usecase, when we are displaying it in code mirror.