svelte: Make compatible with keyword search; remove structural and smart search buttons (#60054)
* svelte: Make compatible with keyword search; remove structural and smart search buttons

This commit aligns the prototype more closely with the production app.
It removes the structural and smart search buttons defaults to keyword
search (which can still be overwritten by server settings).

* Fix lazy loading

It seems the intersecting event wasn't triggered reliably so not all
results loaded. This moves the intersection listener to the last element
of the list.