keyword search: implicit operators and parens as groups
We change the behavior of the parser for keyword search to support the
following:
- treat parenthesis as groups instead of as literals

Before: (a b) -> "(a b)"
Now: (a b) -> (and a b)

- mixing of implicit and explicit boolean operators.

Before: (a or b) c -> ERROR
Now: (a or b) c -> (and (or a b) c)

However, for code search, parenthesis are frequently part of a pattern.
To support searching for functions without escaping the query, we
heuristically parse parenthesis inside pattern as literals.

Before and now: func(a int) -> "func(a int)"
Before and now: () => {} -> (and "()" "=>" "{}")

Test plan:
- Apart from a minor change to the parsing flags, the changes are a NOP for
the other pattern types
- I added a lot of new tests for keyword search. The tests should give a
good overview of the new behavior.