Search jobs: support boolean patterns (#60067)
This PR adds support for boolean patterns to search jobs. It should be okay in
terms of performance, since we recently updated searcher to handle AND/ OR
patterns natively (https://github.com/sourcegraph/sourcegraph/issues/59038).

This also makes the integration with keyword search a lot nicer. Now when you
have a keyword search with multiple terms and you go to create a search job, it
succeeds. Previously it showed a confusing error.