highlighter: De-couple server code from highlighting (#60039)
- `*Query` types are now moved to the server code.
- syntax-analysis crate no longer depends on Rocket (web server).
  Instead, it just takes a FileInfo struct for highlighting,
  and the server code converts a query into a FileInfo value.
- syntax-analysis crate doesn't depend on JSON for highlighting.
  There is still JSON usage in the ctags bit, but we can
  separate that out in a different PR as this is already quite big.

As part of this, I've introduced different types for the different
grammar names (Syntect/Sublime vs Tree-sitter) so that the code is
easier to follow. For simplicity, we do not eagerly validate
the grammar names on construction, but return an error as part of
highlighting if there is no match found. This matches the old
behavior. In the future, we could clean this up to be more
'correct by construction'.