gomod: update or vendor buildkit, docker, hostmatcher and saml to resolve CVEs (#60130)
We had multiple CVEs reported for these dependencies. I don't think this
affected us in practice, but this is a step towards a clean scan from
trivy. I updated to the minimum version which supports the fix.

  go get github.com/moby/buildkit@v0.12.5
  go get github.com/docker/docker@v24.0.7
  go get -u github.com/crewjam/saml

In the case of code.gitea.io/gitea@v1.18.0/modules/hostmatcher we
couldn't update it due to lots of issues popping up in random
transitive dependencies. However, we don't depend on the whole gitea
project, rather just a tiny self contained package in it. So we vendor
it in.

Test Plan: CI and "trivy fs go.mod" reporting no issues.