keyword search: implicit operators and parens as groups (#60157)
We change the behavior of the parser for **keyword search** to support the
following:
- parse parenthesis as groups instead of as literals
- allow mixing of implicit and explicit boolean operators

|query | parser (before) | parser (now) |
|:-----|:-------|:-------|
(a b)| "(a b)" | (and a b) | 
(a or b) c | ERROR | (and (or a b) c)
func(a int)*| "func(a int)" | "func(a int)"

*To support searching for function signatures without escaping the query,
we heuristically parse parenthesis inside patterns as literals. 
We already supported this behavior with "standard" search.
However this heuristic causes **inconsistencies** which may or may not be acceptable, 
depending on how much we want to support queries such as "func(a int)" without quotation.

(foo bar)bas -> (and "foo" "bar" "bas")
foo(bar bas) -> "foo(bar bas)"

Note the lack of space before and after the parenthesis.

Test plan:
- Apart from a minor change to the parsing flags, the changes are a NOP for
the other pattern types
- I added a lot of unit tests for keyword search. The tests should give a
good overview of the new behavior.