code nav: Respect history.preferAbsoluteTimestamps user setting
Closes #51209

This commit changes the logic in the file list panel to respect the
`history.preferAbsoluteTimestamps` setting.

I also noted that `GitCommitNode` is used in most places but the
callsites do not provide the value from the setting. Rather than make
every callsite read the user settings the component itself uses settings
as a fallback when the prop is not explicitly set.