Keyword search: add 'beta' labels and fix URL parsing (#60189)
This PR adds 'Beta' labels to the keyword search banner and popover, as we've
decided to ship the feature as 'beta'.

It also sneaks in a bug fix: before, if the URL did not contain a `patterntype`
parameter, we always defaulted to `standard`. Now, we use the correct default
based on the settings. This is important because users often configure a custom
search engine to search Sourcegraph (so for example they can type
`sg "some query"` in the chrome search bar). Now, these searches use `keyword`
search when the feature is enabled, whereas before they always used `standard`.

(cherry picked from commit cfe547f544d6cb53ad10bdf18ed4b19882e8e447)