gitserver: fix cmd/gitserver/internal/patch.go to use new perforce.NewBaseCommand utility (#60001)
This PR refactors all the "p4" invocations in cmd/gitserver/internal/patch.go to use the new `perforce.NewBaseCommand` API introduced in https://github.com/sourcegraph/sourcegraph/pull/59997. 

As a note,  I think the flow of original code is a bit odd  (a lot of bespoke functions with only one usage), but I believe it's best to keep this change as minimal as possible. 

## Test plan

CI