search: introduce max-line-len streaming parameter
max-line-len is a new parameter which when set will truncate the Content
field of ChunkMatch such that no line is longer than max-line-len.
Currently we only set it for interactive queries in the webapp with the
conservative number 5kb.

A new field is introduced to ChunkMatch when its content has been
truncated. The webapp detects this and will skip syntax highlighting
files which contain truncated ChunkMatches. The webapp will still
highlight the non-truncated ranges.

We have noticed that some documents are close to our max index
size (about 1mb), and are all 1 line. This can lead to response sizes in
the 100s of megabytes. This both is slow for the browser to receive and
parse, as well as causing locking up/freezing of the browser.

The name max-line-len was chosen as a trade-off between brevity and
exactness. I like short key names for URL params that are nearly always
set.

Two alternative approaches were considered. The first is something along
the lines of display limit which omits large events. This had the
downside of potentially displaying nothing or having hard to understand
semantics to avoid that. The second approach was naive truncation. The
downside here is you miss a lot of potential ranges you could highlight
and not having per line information lead to results looking broken.

Test Plan: In my dev environment I:
- lowered the max-line-len to 8
- introduced console logs for skipping syntax highlighting
- introduced an artifical delay to syntax highlighting

The delay was not as important with the final approach I ended up with,
but was kept in testing due to its usefulness to notice how the app
behaves when that is slow.

I then observed that highlighting on the non-truncated parts still
worked. There is one minor artifact of highlighting the text
"...truncated", but that seems acceptable.