gitserver: Fix goroutine leak when reader is not drained (#60238)
When not all messages are read from a reader (for example, if an archive operation times out, or if a ReadFile request is canceled, or the reader is only interested in the first X lines), we didn't properly Wait() on the command.

The added test case proves that before the code changes to the reader, there is 1 goroutine too much still active in the runtime.
After this change, it's equal, across a few hundred runs.

## Test plan

Added test case.

(cherry picked from commit 5de0b7c7ae353d332095b1c090acb24ebc06d4fb)