Fixup and improve Blame error handling (#60108)
While investigating why our blob page would not render a nicer error message, I realized that we threw an error where we shouldn't. Then, a full can of worms opened up. So I went out and tried to fix it.

What's fixed here:

- Error reporting for blob not found is now correctly showing the right component again

**Before**

![Screenshot 2024-02-02 at 19 47 24@2x](https://github.com/sourcegraph/sourcegraph/assets/19534377/7c1e8551-bffd-4968-a9bf-5905c6b830ba)

**After**

![Screenshot 2024-02-02 at 19 45 15@2x](https://github.com/sourcegraph/sourcegraph/assets/19534377/88f51887-94d9-45b8-9e81-4c46e7f326c4)

- When the file does not exist, the Blame endpoint will be called exactly once, and receives a 404 error. 

![Screenshot 2024-02-02 at 19 47 51@2x](https://github.com/sourcegraph/sourcegraph/assets/19534377/7a9e0c03-6b8e-4422-b082-061a32edbda5)

Previously, it would return a 500 error because we had no handling in place that correctly maps to `os.ErrNotExist`. Also, the error handling in the frontend was not working correctly. We now correctly report the error, treat it as non-retriable, and also render the error. 

Before: 
![Screenshot 2024-02-02 at 19 49 55@2x](https://github.com/sourcegraph/sourcegraph/assets/19534377/2dd74a69-573c-4c62-8816-9fd1cd7c3971)


## Test plan

Added test suites, and manually verified what is shown above.