code nav: Respect history.preferAbsoluteTimestamps user setting (#60204)
Closes #51209

This commit changes the logic in the file list panel to respect the
`history.preferAbsoluteTimestamps` setting.

I also noted that `GitCommitNode` is used in most places but the
callsites do not provide the value from the setting. Rather than make
every callsite read the user settings the component itself uses settings
as a fallback when the prop is not explicitly set.

(cherry picked from commit f2448ba6901d25d426cabc701451724332de990f)