gitserver: Reduce frequency of janitor runs and slow start
The frequency of this janitor was WAY too frequent. It's been like that for 5 years, but there's no good reason to run those git maintenance tasks more often than every day.
They're meant to optimize the repository, and spending basically every minute a ton of compute on optimizing is probably not actually making things better.

Test plan:

On sourcegraph.com, this will not make a difference, as the janitor doesn't get through all the repos before the usual 24h restart cycle anyways. It only finishes on weekends, indicating that even a week would probably not be problematic here.
We'll see on S2 that this change will not negatively impact availability.