code-search: handle mobile responsiveness for cody banner on search page (#60301)
After doing some testing on DotCom, I noticed the new cody banner was unresponsive on mobile. This PR fixes that.

I also noticed the `Explore Cody` link wasn't linking to the marketing page on Dotcom. @taiyab I'm not sure if this is better or we should link to the PLG dashboard instead, however, we'll still need to handle when a user isn't authenticated on DotCom - linking to the marketing page `sourcegraph.com/cody` was the easiest of all options, however, I'm happy to change to something else.

# Before

https://github.com/sourcegraph/sourcegraph/assets/25608335/9c050e69-de16-4a85-a4ba-0ec4a5cc96af

# After

https://github.com/sourcegraph/sourcegraph/assets/25608335/248aecea-7699-4a24-aee8-03e50b2962d6

## Test plan

* Manual testing based on the description above.