svelte: Improve scroll behavior on repo pages (#60307)
This fixes an issue that had been bugging for quite a while: When you scrolled within a file and navigated to a new file, we would open the new file at the same scroll position instead of at the top of the file.

SvelteKit has native support for restoring and resetting the scroll position, but that only works when the whole page is scrollable. In our case we made an "app" wrapper <div> scrollable, so that wouldn't work with SvelteKit.
I made the necessary HTML changes so that we can leverage SvelteKit's default behavior but that was problematic too: After clicking on an entry in the file tree the whole file tree would be moved down because now the whole page would be scrolled to the top, revealing the page navigation bar and repo header again.

So what we really want is to scroll to the top of what I consider the content of the repo page, excluding page navigation and repo header. That's exactly what the changes in the repo layout file do: After navigation we check the scroll position of the page (easy to do now that truly the whole page is scrollable), and scroll to the beginning of the page content. We need to take some additional steps to restore the correct scroll position when navigating back and forth via the built-in snapshot feature.