all: update code paths as if qdrant is disabled
Fully removing our use of qdrant is quite a large change. Instead this
does a minimal change by just removing the part of our code which
optionally searches via qdrant, and additionally the part which uses a
real qdrant client. This effectively makes it so qdrant is always
disabled.

In practice, the "GetEmbeddingsConfig" will only return a configuration
for dotcom. Currently qdrant is disabled on dotcom, so operationally
this is a noop change.

Test Plan: CI and investigating that live environments this is a noop.