code-monitors: handle soft-deleted users
Code monitors from soft-deleted users cause errors in the client. For
example, if an admin opens the code monitors page, they see a "user not
found" error.

With this change we filter code monitors of soft-deleted users. This
means the system effectively behaves as if the user was hard-deleted.

For users that were actually hard-deleted, we already delete their code
monitors, which avoids this issue.

Test plan:
new unit test