grpc: Use native gRPC serve handler in repo-updater
The ServeHTTP API of go-gRPC is considered experimental, not feature complete, and has caused issues for us before.
So, to benefit from long-term stability, we want to migrate all our services to serve gRPC natively on a port.

Repo-updater doesn't serve any other HTTP things anymore, so it's a good candidate to get started with!

The `/healthz` is not used in any of our deployment methods, we call `/healthz`, but on the `debugserver` port, not the main server. So I think it should be safe to remove it.

I made a new package called grpcserver that is a very similar copy of the httpserver package, that wraps the logic to create a listener, spawn the server, and listen to Stop() in the form of a goroutine.BackgroundRoutine hook.
Similarly to HTTP, we try a graceful shutdown first, but if that takes too long, we will aggressively ask the server to shut down.

I decided to not implement a middleware to achieve the same "use internal actor only" thing that we had before, and instead now make the functions themselves make this decision, for less of a magical surprise moment.

## Test plan

Tried locally that requests to repo-updater still work. Verified in tracing that the spans show up correctly.
CI integration tests will also confirm this interconnection works properly for the server image as well.