[Backport 5.3] Docs: update links to point to new site
Docs: update links to point to new site (#60381)

We have a number of docs links in the product that point to the old doc site.

Method:
- Search the repo for `docs.sourcegraph.com`
- Exclude the `doc/` dir, all test fixtures, and `CHANGELOG.md`
- For each, replace `docs.sourcegraph.com` with `sourcegraph.com/docs`
- Navigate to the resulting URL ensuring it's not a dead link, updating the URL if necessary

Many of the URLs updated are just comments, but since I'm doing a manual audit of each URL anyways, I felt it was worth it to update these while I was at it.

(cherry picked from commit 1ead9452677f1cf371cff82eed81cdc07a797c33)