release: half fix bake*ExecutorsSteps (#60470)
Even though this code was added a year ago, it clearly has never
actually been run. The first fix is correctly specifying the name of the
terraform-aws-executors repository. It had the name and owner mixed up,
causing the clone to fail.

The second fix is to fetch the history so that prepare-release.sh
actually runs.

Note: this will still just fail. But it gets far along enough in my
release now that I can pick up the pieces and manually create PRs. Then
I can comment out these steps and run the tool again.

Test Plan: blood, sweat and tears