release: skip updating src-cli docs reference (#60480)
I am running the 5.3 release but this is all borked. We used to run a
src-cli reference doc generator, but we now have a docs repo which uses
mdx files. So the reference generator needs to be updated. So for now I
am just skipping this and will follow up later.

Test Plan: progress in running bake-content for release