gitserver: Remove ResolveRevisions from client
To migrate ResolveRevision to gRPC, I'm removing this endpoint in favor of individual calls
for now.
This will simplify the migration. If someone was requesting hundreds of revs here, this will
actually give us better metrics about that, and we'll be able to fix it again.

Test plan:

Adjusted existing tests.