search: better handling of keyword boundaries (#60477)
This fixes a bug in the backend parser to recognise "AND" and "OR" even if they aren't preceded by whitespace. Previously, a query such as "(a or b)and c" was not parsed correctly, because "and" was preceded by ")", which resulted in surprising results.

They key idea of the fix is to recognise parenthesis as proper boundaries and to not advance the parser immediately once we encounter a right parenthesis, but instead to let the caller advance. This way, the caller can construct an operator (`(or a b)` in this case) before continuing with the parsing. This mirrors the behavior of the other keywords that can break the loop in the leaf parser.

The handling of whitespace in our parser is very fragile. A more thorough fix would be to introduce a lexer to eliminate the significance of whitespace from the parser logic. However, we are planning to replace the parser soon. Right now it makes sense to fix this bug within the current framework.

## Test plan
- new units tests
- all existing test pass