gitserver: Calculate repo size less frequently in janitor (#60499)
In the janitor, we calculate the repo size. That is quite IO heavy and expensive, and every fetch (basically the only writing operation) runs this, too.
So janitor is mostly doing that to ensure that things don't get terribly out of sync, if some issues occur.

Given that, we don't need to run this every time. Instead, we can run it only once a day. In reality, probably less often. We can adjust as needed.

Also moves it to the end of the janitor operations, to collect a size that's not outdated immediately when the next janitor step runs.

## Test plan

Verified the setting works locally.