search: replace autogold with table based test
This is the first PR in a series to replace autogold with simpler table
based tests where it makes sense.

Why?

Autogold is great for regression tests or when the expected result is
tedious to write out, but it is more difficult to parse visually and it
is not as convenient during debugging.

Test plan:
all tests pass