gitserver: Remove effectively disabled debouncing of updates
The only value we ever passed here was 1 second, which is effectively "don't debounce".
So we can simplify here.

Really, the client should make sure to go at a reasonable pace, and gitserver on the
other end of the wire can have ultimate authority over if it's time for an update or
not.

All existing tests still pass, integration and E2E tests are passing.