conf: Allow a local stub file
See the inline comments, this PR will allow to add standalone commands to `sg` where we can run e.g. gitserver without requiring a frontend to be running.
This can be useful to test services without the full stack, which reduces load and recompile times, and doesn't trigger any requests to the services that you don't want.

## Test plan

Wrote a standalone config for gitserver and it can start up and run properly.