gitserver: Reduce "normalization" of repo names
This normalization might've been useful 5 years ago when gitserver
tried to clone whatever it was given, but since then we introduced
the DB layer where we have to go through the GetRemoteURL func.
That one verifies that a repo is meant to be cloned into the sg instance,
but it ALSO requires that the repo name is a LITERAL match. When we make
modifications to the repo name, it will no longer match.
That also applies to various of our gitserver_repos queries, which are all
depending on LITERAL name matches.

Ideally, we remove this altogether, and instead use a "verify the path is
safe" only.