grpc: Bump streamio code to latest version
As I dug into why we're allocating a ton of memory in io.Copy when forwarding the stdout from exec calls to the writer, I found that we allocate 128kb every time, while many calls actually only write 40 bytes (a commit SHA).

I then looked at gitaly and noticed that we use a very old version of their implementation.

So here we go, updating it to the latest.
Note that we currently don't need the NewSyncWriter, but I kept it anyways, maybe it reminds us of the importance if we ever need it.

List of commits from gitaly: https://gitlab.com/gitlab-org/gitaly/-/commits/master/streamio/stream.go?ref_type=heads

---

streamio: Implement synchronized writer
Writing to gRPC streams concurrently isn't allowed, which is why we need
to synchronize those writes. Doing so with the `NewWriter()` is
repetitive, so this commit implements a new function `NewSyncWrite()`
which receives a mutex in addition to the callback function. The mutex
is then locked and unlocked previous to invoking the callback.

---

streamio: Simplify implementation of ReadFrom
The ReadFrom implementation is quite hard to read as the loop depends on
state from the previous iteration. This commit refactors the code to not
do so anymore by always sending data if we've read something, even if
there was a read error.

---

streamio: Clarify why read errors are deferred
The implementation of `Read()` only returns read errors in case no data
is left in the buffer. This is required such that we correctly drain any
buffered data and only return the error when no data is left. This may
not be immediately clear, so this commit puts a comment on this and adds
a testcase which exercises this code path.

---

streamio: remove custom ReadFrom and WriteTo
This replaces sendWriter.ReadFrom and receiveReader.WriteTo with
trivial implementations based on io.Copy, to be removed in 14.0.

The purpose of the io.ReaderFrom and io.WriterTo interfaces is to
optimize away a buffer allocation in io.Copy, and/or being able to do
something smart like calling sendfile(2). Without them, io.Copy
allocates a 32KB buffer and reads into / writes from that buffer in a
loop.

The thing is, sendWriter.ReadFrom and receiveReader.WriteTo neither
make a difference for allocations, nor do they do something "smart".

First of all, sendWriter.ReadFrom was not helping the buffer
allocation efficiency because it allocated its own 128KB buffer. In
fact, sendWriter.ReadFrom is equivalent to what io.Copy does on the
inside, so we might as well let io.Copy do that.

That leaves receiveReader.WriteTo as a potentially useful
optimization. It does prevent io.Copy from allocating that 32KB
buffer. However, everytime it calls receiver(), gRPC allocat...

---

Remove streamio ReaderFrom and WriterTo
These were set to be removed in v14.

Changelog: removed

---

streamio.Reader: remember errors
If there has been an error, subsequent calls to Read should keep
returning that error.

Changelog: fixed

## Test plan

Verified locally that sourcegraph still works for basic clickops. CI will verify that archive fetching etc still behaves as previously.
I will take another look at the cloud profiler once this has been rolled out.